+++ test/externalTests/common.sh
@@ -173,7 +173,7 @@ function run_install
 
     replace_version_pragmas
     force_truffle_solc_modules "$soljson"
-    force_truffle_compiler_settings "$CONFIG" "${DIR}/solc" "$OPTIMIZER_LEVEL" istanbul
+    force_truffle_compiler_settings "$CONFIG" "${DIR}/solc" "$OPTIMIZER_LEVEL" berlin
 
     $init_fn
 }
@@ -234,7 +234,7 @@ function truffle_run_test
     for level in $(seq "$OPTIMIZER_LEVEL" 3)
     do
         clean
-        force_truffle_compiler_settings "$CONFIG" "${DIR}/solc" "$level" istanbul
+        force_truffle_compiler_settings "$CONFIG" "${DIR}/solc" "$level" berlin
 
         printLog "Running compile function..."
         $compile_fn
@@ -267,4 +267,3 @@ function external_test
     rm -rf "$DIR"
     echo "Done."
 }
-